requirements.txt
flask>=3.0.0
requests>=2.31.0
python-dotenv>=1.0.0
transformers>=4.35.0
torch>=2.0.0
spotipy>=2.22.0
lyricsgenius>=3.0.0
langchain>=0.1.0
langchain-community>=0.0.6
openai>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
beautifulsoup4>=4.12.0
streamlit>=1.28.0
plotly>=5.0.0
